refactor: rename boardForm to ModalForm and move it to src folder
